Update shard dependencies.
Specify just the minimal version of dependencies that are expected to work, so
we don't lock the version of anyone that also depends on these libs in their
projects.

In case of some future version of these dependencies don't work with sidekiq, a
bug will be filledand we can fix it, so everybody will be always happy.